+++ crates/schematic/src/config/mod.rs
@@ -27,3 +27,8 @@ macro_rules! derive_enum {
         $impl
     };
 }
+
+pub type DefaultValueResult<T> = std::result::Result<Option<T>, HandlerError>;
+pub type ParseEnvResult<T> = std::result::Result<Option<T>, HandlerError>;
+pub type MergeResult<T> = std::result::Result<Option<T>, HandlerError>;
+pub type ValidateResult = std::result::Result<(), ValidateError>;