+++ crates/schematic/src/env.rs
@@ -1,9 +1,8 @@
-use crate::config::ConfigError;
-use crate::internal;
+use crate::{internal, ParseEnvResult};
 use std::str::FromStr;
 
 /// Ignore the environment variable if it's empty and fallback to the previous or default value.
-pub fn ignore_empty<T: FromStr>(var: String) -> Result<Option<T>, ConfigError> {
+pub fn ignore_empty<T: FromStr>(var: String) -> ParseEnvResult<T> {
     let var = var.trim();
 
     if var.is_empty() {
@@ -15,14 +14,14 @@ pub fn ignore_empty<T: FromStr>(var: String) -> Result<Option<T>, ConfigError> {
 
 /// Parse a string into a boolean. Will parse `1`, `true`, `yes`, `on`,
 /// and `enabled` as true, and everything else as false.
-pub fn parse_bool(var: String) -> Result<Option<bool>, ConfigError> {
+pub fn parse_bool(var: String) -> ParseEnvResult<bool> {
     Ok(match var.to_lowercase().as_str() {
         "1" | "true" | "yes" | "on" | "enabled" | "enable" => Some(true),
         _ => Some(false),
     })
 }
 
-fn split<T: FromStr>(var: String, delim: char) -> Result<Option<Vec<T>>, ConfigError> {
+fn split<T: FromStr>(var: String, delim: char) -> ParseEnvResult<Vec<T>> {
     let mut list = vec![];
 
     for s in var.split(delim) {
@@ -37,21 +36,21 @@ fn split<T: FromStr>(var: String, delim: char) -> Result<Option<Vec<T>>, ConfigE
 }
 
 /// Split a variable on each comma (`,`) and parse into a list of values.
-pub fn split_comma<T: FromStr>(var: String) -> Result<Option<Vec<T>>, ConfigError> {
+pub fn split_comma<T: FromStr>(var: String) -> ParseEnvResult<Vec<T>> {
     split(var, ',')
 }
 
 /// Split a variable on each colon (`:`) and parse into a list of values.
-pub fn split_colon<T: FromStr>(var: String) -> Result<Option<Vec<T>>, ConfigError> {
+pub fn split_colon<T: FromStr>(var: String) -> ParseEnvResult<Vec<T>> {
     split(var, ':')
 }
 
 /// Split a variable on each semicolon (`;`) and parse into a list of values.
-pub fn split_semicolon<T: FromStr>(var: String) -> Result<Option<Vec<T>>, ConfigError> {
+pub fn split_semicolon<T: FromStr>(var: String) -> ParseEnvResult<Vec<T>> {
     split(var, ';')
 }
 
 /// Split a variable on each space (` `) and parse into a list of values.
-pub fn split_space<T: FromStr>(var: String) -> Result<Option<Vec<T>>, ConfigError> {
+pub fn split_space<T: FromStr>(var: String) -> ParseEnvResult<Vec<T>> {
     split(var, ' ')
 }