+++ crates/schematic/src/merge.rs
@@ -1,41 +1,40 @@
-use crate::config::{ConfigError, PartialConfig};
+use crate::{
+    config::{HandlerError, PartialConfig},
+    MergeResult,
+};
 use std::{
     collections::{BTreeMap, BTreeSet, HashMap, HashSet},
     hash::Hash,
 };
 
 /// A merger function that receives the previous and next values, the current
-/// context, and can return a [`ConfigError`] on failure.
-pub type Merger<Val, Ctx> = Box<dyn FnOnce(Val, Val, &Ctx) -> Result<(), ConfigError>>;
+/// context, and can return a [`HandlerError`] on failure.
+pub type Merger<Val, Ctx> = Box<dyn FnOnce(Val, Val, &Ctx) -> MergeResult<Val>>;
 
 /// Discard both previous and next values and return [`None`].
-pub fn discard<T, C>(_: T, _: T, _: &C) -> Result<Option<T>, ConfigError> {
+pub fn discard<T, C>(_: T, _: T, _: &C) -> MergeResult<T> {
     Ok(None)
 }
 
 /// Always preserve the previous value over the next value.
-pub fn preserve<T, C>(prev: T, _: T, _: &C) -> Result<Option<T>, ConfigError> {
+pub fn preserve<T, C>(prev: T, _: T, _: &C) -> MergeResult<T> {
     Ok(Some(prev))
 }
 
 /// Always replace the previous value with the next value.
-pub fn replace<T, C>(_: T, next: T, _: &C) -> Result<Option<T>, ConfigError> {
+pub fn replace<T, C>(_: T, next: T, _: &C) -> MergeResult<T> {
     Ok(Some(next))
 }
 
 /// Append the items from the next vector to the end of the previous vector.
-pub fn append_vec<T, C>(
-    mut prev: Vec<T>,
-    next: Vec<T>,
-    _: &C,
-) -> Result<Option<Vec<T>>, ConfigError> {
+pub fn append_vec<T, C>(mut prev: Vec<T>, next: Vec<T>, _: &C) -> MergeResult<Vec<T>> {
     prev.extend(next);
 
     Ok(Some(prev))
 }
 
 /// Prepend the items from the next vector to the start of the previous vector.
-pub fn prepend_vec<T, C>(prev: Vec<T>, next: Vec<T>, _: &C) -> Result<Option<Vec<T>>, ConfigError> {
+pub fn prepend_vec<T, C>(prev: Vec<T>, next: Vec<T>, _: &C) -> MergeResult<Vec<T>> {
     let mut new = vec![];
     new.extend(next);
     new.extend(prev);
@@ -49,7 +48,7 @@ pub fn merge_btreemap<K, V, C>(
     mut prev: BTreeMap<K, V>,
     next: BTreeMap<K, V>,
     _: &C,
-) -> Result<Option<BTreeMap<K, V>>, ConfigError>
+) -> MergeResult<BTreeMap<K, V>>
 where
     K: Eq + Hash + Ord,
 {
@@ -65,7 +64,7 @@ pub fn merge_btreeset<T, C>(
     mut prev: BTreeSet<T>,
     next: BTreeSet<T>,
     _: &C,
-) -> Result<Option<BTreeSet<T>>, ConfigError>
+) -> MergeResult<BTreeSet<T>>
 where
     T: Eq + Hash + Ord,
 {
@@ -82,7 +81,7 @@ pub fn merge_hashmap<K, V, C>(
     mut prev: HashMap<K, V>,
     next: HashMap<K, V>,
     _: &C,
-) -> Result<Option<HashMap<K, V>>, ConfigError>
+) -> MergeResult<HashMap<K, V>>
 where
     K: Eq + Hash,
 {
@@ -94,11 +93,7 @@ where
 }
 
 /// Shallow merge the next [`HashSet`] into the previous [`HashSet`], overwriting duplicates.
-pub fn merge_hashset<T, C>(
-    mut prev: HashSet<T>,
-    next: HashSet<T>,
-    _: &C,
-) -> Result<Option<HashSet<T>>, ConfigError>
+pub fn merge_hashset<T, C>(mut prev: HashSet<T>, next: HashSet<T>, _: &C) -> MergeResult<HashSet<T>>
 where
     T: Eq + Hash,
 {
@@ -115,8 +110,9 @@ pub fn merge_partial<T: PartialConfig>(
     mut prev: T,
     next: T,
     context: &T::Context,
-) -> Result<Option<T>, ConfigError> {
-    prev.merge(context, next)?;
+) -> MergeResult<T> {
+    prev.merge(context, next)
+        .map_err(|error| HandlerError(error.to_string()))?;
 
     Ok(Some(prev))
 }